move all library code to lib module and expose default functions through root module
